Module ripalt::api::identity

Request identity service for Actix applications.

IdentityService middleware can be used with different policies types to store identity information.

By default, only the api identity policy is implemented. Other backend implementations can be added separately.

ApiIdentityPolicy uses the session or a JWT as identity storage.

To access current request identity RequestIdentity should be used. HttpRequest implements RequestIdentity trait.

use actix_web::middleware::identity::RequestIdentity;
use actix_web::*;
use ripalt::api::identity::{ApiIdentityPolicy, IdentityService};

fn index(req: HttpRequest) -> Result<String> {
    // access request identity
    if let Some(id) = req.identity() {
        Ok(format!("Welcome! {}", id))
    } else {
        Ok("Welcome Anonymous!".to_owned())
    }
}

fn main() {
    let app = App::new().middleware(IdentityService::new(
        // <- create identity middleware
        ApiIdentityPolicy::new(&[0; 32])    // <- create api session backend
    ));
}

Structs

ApiIdentity
Identity that uses the Session or a JWT as identity storage.

ApiIdentityPolicy
The API identity policy

Claims
The claims for the JWT

IdentityService
Request identity middleware

Traits

Identity
An identity

IdentityPolicy
Identity policy definition.

RequestIdentity
The helper trait to obtain your identity from a request.