Trait ripalt::api::identity::RequestIdentity

pub trait RequestIdentity {
    fn identity(&self) -> Option<&str>;

    fn forget(&mut self);

    fn credentials(&self) -> Option<(&Uuid, &Uuid)>;

    fn user_id(&self) -> Option<&Uuid> { ... }

    fn group_id(&self) -> Option<&Uuid> { ... }
}

The helper trait to obtain your identity from a request.

use actix_web::*;
use ripalt::api::identity::RequestIdentity;

fn index(req: HttpRequest) -> Result<String> {
    // access request identity
    if let Some(id) = req.identity() {
        Ok(format!("Welcome! {}", id))
    } else {
        Ok("Welcome Anonymous!".to_owned())
    }
}

Required Methods

fn identity(&self) -> Option<&str>

Return the claimed identity of the user associated request or None if no identity can be found associated with the request.

fn forget(&mut self)

This method is used to 'forget' the current identity on subsequent requests.

fn credentials(&self) -> Option<(&Uuid, &Uuid)>

Provided Methods

fn user_id(&self) -> Option<&Uuid>

fn group_id(&self) -> Option<&Uuid>