Struct ripalt::models::acl::Acl

pub struct Acl { /* fields omitted */ }

The ACL Structure

contains all user (AclUserRule) and group rules (AclGroupRule)

Methods

impl Acl

pub fn new() -> Self

Create a new empty Acl object

pub fn load(&mut self, db: &PgConnection)

Load the ACL structure from the database

Panics

This function panics if the underlying database shits the bed.

pub fn reload(&mut self, db: &PgConnection)

Reload the ACL structure from the database

Panics

This function panics if the underlying database shits the bed.

pub fn is_allowed(
&self,
uid: &Uuid,
gid: &Uuid,
ns: &str,
perm: &Permission
) -> bool

Check if the user is allowed to do perm in namespace ns

user - the user to test

ns - acl namespace

perm - permission to test

pub fn is_group_allowed(&self, gid: &Uuid, ns: &str, perm: &Permission) -> bool

Check if the group is allowed to do perm in namespace ns

if no explicit rule for the group is found check the parent group(s) recursively until an explicit rule is found. If no rule is found the function returns false