Struct ripalt::models::chat::ChatMessageWithUser

pub struct ChatMessageWithUser {
    pub id: Uuid,
    pub user_id: Uuid,
    pub chat: i16,
    pub message: String,
    pub created_at: Timestamp,
    pub user_name: String,
    pub user_group: Uuid,
}

Chat message data structure

is the same as ChatMessage, but with the user name and user group added.

Used whenever a chat message needs to be returned to the user.

Fields
id: Uuid

the unique message id

user_id: Uuid

user id

chat: i16

chatroom. Can be converted to ChatRoom.

message: String

the actual message

created_at: Timestamp

timestamp when the message was created

user_name: String

user name

user_group: Uuid

user group

Methods

impl ChatMessageWithUser

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn messages_for_chat(
chat: i16,
since: Option<Timestamp>,
limit: i64,
db: &PgConnection
) -> Vec<Self>

Fetch all messages for a chatroom

Filters by chat.

If since is Some Timestamp, it only returns message newer than the timestamp.

Returns at most limit messages