Struct ripalt::models::user::User

pub struct User {
    pub id: Uuid,
    pub name: String,
    pub email: String,
    pub password: Bytes,
    pub salt: Bytes,
    pub status: i16,
    pub created_at: Timestamp,
    pub updated_at: Timestamp,
    pub passcode: Bytes,
    pub uploaded: i64,
    pub downloaded: i64,
    pub group_id: Uuid,
    pub ip_address: Option<IpNetwork>,
    pub last_active: Option<Timestamp>,
}

Fields
id: Uuid name: String email: String password: Bytes salt: Bytes status: i16 created_at: Timestamp updated_at: Timestamp passcode: Bytes uploaded: i64 downloaded: i64 group_id: Uuid ip_address: Option<IpNetwork> last_active: Option<Timestamp>
Methods

impl User

pub fn find(id: &Uuid, db: &PgConnection) -> Option<User>

Find an User by its ID

pub fn find_by_name(name: &str, db: &PgConnection) -> Option<User>

Find an User by the username

pub fn find_by_email(email: &str, db: &PgConnection) -> Option<User>

Find an User by the email address

pub fn find_by_passcode(passcode: &[u8], db: &PgConnection) -> Option<User>

Find an User by the passcode

pub fn set_password(&mut self, plain_password: &str)

Set the password to the generated hash of the given password.

Generates a new SALTBYTES byte salt and sets self.salt to this salt.

See the util::password module for the hashing implementation.

pub fn verify_password(&self, password: &str) -> bool

Verify that he given password is identical to the stored one

See the util::password module for the verify implementation.

pub fn create(
db: &PgConnection,
name: String,
email: String,
password: &str,
group: &Group
) -> Result<User, Error>

Create a new User

pub fn create_confirm_id(&self, db: &PgConnection) -> Bytes

Create a new confirm id

Returns

the confirm id

pub fn save(&self, db: &PgConnection) -> Result<usize, Error>

Save the User into the database

pub fn update_last_active(&mut self, db: &PgConnection) -> Result<usize, Error>

Update the last_active timestamp to now.

pub fn create_message_folders(&self, db: &PgConnection) -> Result<(), Error>

Create the Message Folders