Struct ripalt::template::Template

pub struct Template {
    pub body: String,
    pub content_type: String,
}

Container for the rendered template data

Template can be used as a Responder or HttpResponse for handler functions.

Fields
body: String content_type: String
Methods

impl Template

pub fn new(body: String, content_type: String) -> Self

Create a new Template container

pub fn render<T>(
tpl: &TemplateSystem,
name: &str,
ctx: T
) -> Result<HttpResponse> where
T: Serialize,

Render a registered template

render returns a Result, which is suitable to be returned by a handler function. If the template fails to render, Error is set to ErrorInternalServerError